grub.cfg.main: Set default rootmode=ro for kernel mount
Pass rootmode=ro the to kernel bootargs.
This is the Linux default approach of mounting the rootfs.
If rootfs mount mode equal to “ro” and rootfs flag is “dirty”, then the
fsck utility is running automatically and fixes fs errors. Overwise, if
rootmode!=ro, then this logics get not invoked.
TBD: fsck=force fsck.repair=yes

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>