layer.conf: Fix WKS_BOOT_SIZE value
The WKS_BOOT_SIZE variable is used in wic/imx-efi-root-boot-bootpart.wks.in
as the value for the '--size' parameter in the partition definition:

    part /boot --source rootfs --rootfs-dir=${IMAGE_ROOTFS}/boot \
         --ondisk mmcblk --label boot --active --align 8192 \
         --uuid ${PARTUUID_BOOT} --fsuuid ${UUID_BOOT} \
         --size ${WKS_BOOT_SIZE}

According to the wic kickstart documentation, the '--size' parameter
expects a numeric value in megabytes by default when no unit suffix
is specified. Providing "128M" would be interpreted incorrectly,
while "128" is properly interpreted as 128 megabytes.

This change ensures the boot partition size is correctly configured
as 128 MB.

Signed-off-by: Piotr Łobacz <piotr@fever.energy>